NOCOMMIT: bpf: Add bpf_rbtree_node_xchg helper
This is a kludge to enable rbtree node stashing. When this is converted
to kptr style datastructure ~all of this can be deleted, as evidenced by
the strong resemblance of this logic to existing kptr xchg logic.

Needed to add some BTF checking to determine whether a map value BTF
type is a pointer to a rbtree_node, since only maps with those values
should free the rbtree node their values are pointing to.

Tests / example in separate commit which should immediately follow this
one.

Signed-off-by: Dave Marchevsky <davemarchevsky@fb.com>
(cherry picked from commit d5f707b56134028c41293dbe40e7450cdd851b7a)